+++ dataflow/dags/ons_pipelines.py
@@ -52,7 +52,7 @@ class ONSUKSATradeInGoodsPollingPipeline(_FastPollingPipeline):
 
 
 class ONSUKTradeInGoodsByCountryAndCommodityPollingPipeline(_FastPollingPipeline):
-    from dataflow.ons_scripts.uktradecountrybycommodity.main import (
+    from dataflow.ons_scripts.uk_trade_country_by_commodity.main import (
         get_current_and_next_release_date,
         get_data,
     )
@@ -78,3 +78,62 @@ class ONSUKTradeInGoodsByCountryAndCommodityPollingPipeline(_FastPollingPipeline
         ],
     )
     worker_queue = 'high-memory-usage'
+
+
+class ONSUKTradeInServicesByPartnerCountryNSAPollingPipeline(_FastPollingPipeline):
+    from dataflow.ons_scripts.uk_trade_in_services_service_type_by_partner_country_non_seasonally_adjusted.main import (
+        get_current_and_next_release_date,
+        get_data,
+    )
+
+    schedule_interval = "0 6 20-31 1,4,7,10 *"
+
+    date_checker = get_current_and_next_release_date
+    data_getter = get_data
+
+    table_config = SingleTableConfig(
+        schema='ons',
+        table_name='uk_trade_in_services_by_country_nsa',
+        field_mapping=[
+            ('ons_iso_alpha_2_code', sa.Column('ons_iso_alpha_2_code', sa.Text)),
+            ('ons_region_name', sa.Column('ons_region_name', sa.Text)),
+            ('period', sa.Column('period', sa.Text)),
+            ('period_type', sa.Column('period_type', sa.Text)),
+            ('direction', sa.Column('direction', sa.Text)),
+            ('product_code', sa.Column('product_code', sa.Text)),
+            ('product_name', sa.Column('product_name', sa.Text)),
+            ('measure_type', sa.Column('measure_type', sa.Text)),
+            ('total', sa.Column('total', sa.Numeric)),
+            ('unit', sa.Column('unit', sa.Text)),
+            ('marker', sa.Column('marker', sa.Text)),
+        ],
+    )
+
+
+class ONSUKTotalTradeAllCountriesNSAPollingPipeline(_FastPollingPipeline):
+    from dataflow.ons_scripts.uk_total_trade_all_countries_non_seasonally_adjusted.main import (
+        get_current_and_next_release_date,
+        get_data,
+    )
+
+    schedule_interval = "0 6 20-31 1,4,7,10 *"
+
+    date_checker = get_current_and_next_release_date
+    data_getter = get_data
+
+    table_config = SingleTableConfig(
+        schema='ons',
+        table_name="uk_total_trade_all_countries_nsa",
+        field_mapping=[
+            ('ons_iso_alpha_2_code', sa.Column('ons_iso_alpha_2_code', sa.Text)),
+            ('ons_region_name', sa.Column('ons_region_name', sa.Text)),
+            ('period', sa.Column('period', sa.Text)),
+            ('period_type', sa.Column('period_type', sa.Text)),
+            ('direction', sa.Column('direction', sa.Text)),
+            ('product_name', sa.Column('product_name', sa.Text)),
+            ('measure_type', sa.Column('measure_type', sa.Text)),
+            ('value', sa.Column('value', sa.Numeric)),
+            ('unit', sa.Column('unit', sa.Text)),
+            ('marker', sa.Column('marker', sa.Text)),
+        ],
+    )