+++ tests/unit/dags/test_ons_pipelines.py
@@ -1,6 +1,8 @@
 from dataflow.dags.ons_pipelines import (
-    ONSUKTradeInGoodsByCountryAndCommodityPollingPipeline,
     ONSUKSATradeInGoodsPollingPipeline,
+    ONSUKTotalTradeAllCountriesNSAPollingPipeline,
+    ONSUKTradeInServicesByPartnerCountryNSAPollingPipeline,
+    ONSUKTradeInGoodsByCountryAndCommodityPollingPipeline,
 )
 from tests.unit.utils import get_polling_dag_tasks
 
@@ -15,6 +17,24 @@ class TestONSUKSATradeInGoodsPollingPipeline:
 class TestONSUKTradeInGoodsByCountryAndCommodity:
     def test_tasks_in_dag(self):
         dag = ONSUKTradeInGoodsByCountryAndCommodityPollingPipeline().get_dag()
+
+        assert {t.task_id for t in dag.tasks} == get_polling_dag_tasks(
+            with_emails=False
+        )
+
+
+class TestONSUKTradeInServicesByPartnerCountryNSAPollingPipeline:
+    def test_tasks_in_dag(self):
+        dag = ONSUKTradeInServicesByPartnerCountryNSAPollingPipeline().get_dag()
+
+        assert {t.task_id for t in dag.tasks} == get_polling_dag_tasks(
+            with_emails=False
+        )
+
+
+class TestONSUKTotalTradeAllCountriesNSAPollingPipeline:
+    def test_tasks_in_dag(self):
+        dag = ONSUKTotalTradeAllCountriesNSAPollingPipeline().get_dag()
 
         assert {t.task_id for t in dag.tasks} == get_polling_dag_tasks(
             with_emails=False