+++ CHANGELOG.md
@@ -15,6 +15,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Contact and advisor last interaction pipelines.
 
+### Changed
+- ONS UK trade in services migrated to new polling pipeline.
+- ONS UK total trade migrated to new polling pipeline.
+
 ## 2020-09-12
 
 ### Added